add permissions for requests, change method for find out event and some logic changes

+++ src/main/java/org/bitebuilders/controller/EventController.java
@@ -1,31 +1,35 @@
 package org.bitebuilders.controller;
 
+import org.bitebuilders.component.UserContext;
 import org.bitebuilders.controller.dto.EventDTO;
 import org.bitebuilders.controller.dto.MessageResponseDTO;
 import org.bitebuilders.controller.requests.EventRequest;
 import org.bitebuilders.model.Event;
+import org.bitebuilders.model.UserInfo;
 import org.bitebuilders.service.UserInfoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.bitebuilders.service.EventService;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/events")
 public class EventController {
 
-    @Autowired
     private final EventService eventService;
 
-    @Autowired
     private final UserInfoService userInfoService;
 
-    public EventController(EventService eventService, UserInfoService userInfoService) {
+    private final UserContext userContext;
+
+    @Autowired
+    public EventController(EventService eventService, UserInfoService userInfoService, UserContext userContext) {
         this.eventService = eventService;
         this.userInfoService = userInfoService;
+        this.userContext = userContext;
     }
 
     @GetMapping("/active")
@@ -34,48 +38,37 @@ public class EventController {
                 .stream()
                 .map(Event::toEventDTO)
                 .toList();
+
         if (activeEvents != null)
             return ResponseEntity.ok(activeEvents);  // Возвращаем список активных событий
+
         return ResponseEntity.noContent().build();
     }
 
-    // 3. Получение всех мероприятий по adminId
-    @GetMapping("/admin/{adminId}")
-    public ResponseEntity<List<EventDTO>> getEventsByAdminId(@PathVariable Long adminId) {
-        List<EventDTO> eventDTOS = eventService.getEventsByAdminId(adminId)
+    @GetMapping("/my")
+    public ResponseEntity<List<EventDTO>> getMyEvents() {
+        UserInfo user = userContext.getCurrentUser();
+        List<EventDTO> eventDTOS = userInfoService.getMyEvents(user)
                 .stream()
                 .map(Event::toEventDTO)
                 .toList();
-        if (eventDTOS != null)
-            return ResponseEntity.ok(eventDTOS);  // Возвращаем список событий для конкретного adminId
-        return ResponseEntity.noContent().build();
+
+        return ResponseEntity.ok(eventDTOS);
     }
 
     // Получение мероприятия по его id
     @GetMapping("/{eventId}")
+    @PreAuthorize("hasRole('MANAGER') or hasRole('ADMIN')")
     public ResponseEntity<EventDTO> getEventByID(@PathVariable Long eventId) {
-        Optional<Event> foundEvent = eventService.getEventById(eventId);
-        return foundEvent
-                .map(event -> ResponseEntity.ok(event.toEventDTO()))
-                .orElse(ResponseEntity.notFound().build());
+        if (!eventService.haveManagerAdminAccess(eventId)) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        Event foundEvent = eventService.getEventById(eventId);
+        return ResponseEntity.ok(foundEvent.toEventDTO());
     }
 
-    // Получение статуса мероприятия по его id
-    @GetMapping("/condition/{eventId}")
-    public ResponseEntity<Event.Condition> getEventConditionByID(@PathVariable Long eventId) {
-        Event.Condition result = eventService.getEventCondition(eventId);
-        return ResponseEntity.ok(result);
-    }
-
-    @GetMapping("/my-events/{userId}")
-    public ResponseEntity<List<EventDTO>> getMyEvents(@PathVariable Long userId) { // TODO без pathvariable будет
-        List<EventDTO> eventDTOS = userInfoService.getMyEvents(userId)
-                .stream()
-                .map(Event::toEventDTO)
-                .toList();
-        return ResponseEntity.ok(eventDTOS);
-    }
-
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/post")
     public ResponseEntity<EventDTO> createEvent(@RequestBody EventRequest requestedEvent) {
         Event newEvent = requestedEvent.toEvent();
@@ -88,47 +81,69 @@ public class EventController {
     /**
      * Запускает мероприятие досрочно
      */
+    @PreAuthorize("hasRole('MANAGER') or hasRole('ADMIN')")
     @PostMapping("/start-event")
     public ResponseEntity<EventDTO> startEvent(@RequestBody Long eventId) {
         Event startedEvent;
+
+        if (!eventService.haveManagerAdminAccess(eventId))
+            return ResponseEntity.badRequest().build();
+
         try {
             startedEvent = eventService.startEventById(eventId);
         } catch (IllegalStateException e) {
             System.err.println(e.getMessage());
             return ResponseEntity.badRequest().build();
         }
+
         return ResponseEntity.ok(startedEvent.toEventDTO());
     }
 
+    @PreAuthorize("hasRole('MANAGER') or hasRole('ADMIN')")
     @PutMapping("/update/{eventId}")
     public ResponseEntity<EventDTO> updateEvent(@PathVariable Long eventId,
                                              @RequestBody EventRequest requestedEvent) {
+        if (!eventService.haveManagerAdminAccess(eventId))
+            return ResponseEntity.badRequest().build();
+
         Event eventToUpdate = requestedEvent.toEvent(eventId);
         EventDTO eventDTO = eventService.createOrUpdateEvent(eventToUpdate).toEventDTO();
+
         if (eventDTO != null)
             return ResponseEntity.ok(eventDTO);
+
         return ResponseEntity.notFound().build();
     }
 
+    @PreAuthorize("hasRole('MANAGER') or hasRole('ADMIN')")
     @PutMapping("/hide/{eventId}")
     public ResponseEntity<MessageResponseDTO> hideEvent(@PathVariable Long eventId) {
-        Boolean result = eventService.hideEvent(eventId);
+        if (!eventService.haveManagerAdminAccess(eventId))
+            return ResponseEntity.badRequest().build();
+
+        boolean result = eventService.hideOrFindOutEvent(eventId);
         if (result) {
             return ResponseEntity.ok(
                     new MessageResponseDTO("Event with id " + eventId + " hided successfully"));
         }
+
         return ResponseEntity.badRequest().body(
                 new MessageResponseDTO("Event with id " + eventId + " could not hide")
         );
     } // TODO вернуть в активное состояние
 
+    @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/delete/{eventId}")
     public ResponseEntity<MessageResponseDTO> deleteEventById(@PathVariable Long eventId) {
-        Boolean result = eventService.deleteEvent(eventId);
+        if (!eventService.haveAdminAccess(eventId))
+            return ResponseEntity.badRequest().build();
+
+        boolean result = eventService.deleteEvent(eventId);
         if (result) {
             return ResponseEntity.ok(
                     new MessageResponseDTO("Event with id " + eventId + " deleted successfully"));
         }
+
         return ResponseEntity.badRequest().body(
                 new MessageResponseDTO("Event with id " + eventId + " could not delete")
         );
